# wwwdir/venom/shared/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Api
  description: VENOM CODES API specification — WHMCS-backed client portal
  version: 0.1.0
servers:
  - url: /
paths:
...
  /api/users/permissions-list:
    get:
      tags: [Users]
      summary: WHMCS GetPermissionsList
      operationId: getPermissionsList
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AnyJson:
      type: object
      additionalProperties: true
...